refactor: Pair identification test moved to SAMRecordUnitTest
